requirements.txt
setuptools==44.0.0
wheel==0.36.2
pylint==2.6.0
autopep8==1.5.4
alpha-vantage==2.2.0    
yfinance==0.1.55   
spacy-nightly==3.0.0rc2